requirements.txt
numpy                     1.24.4                   
opencv-contrib-python     4.10.0.84                
opencv-python             4.10.0.84                
openssl                   3.3.1               
opt-einsum                3.3.0                   
paddleocr                 2.8.1                   
paddlepaddle              2.6.1                    
pandas                    2.0.3                                      
pillow                    10.4.0                   
pip                       24.0            
pytesseract               0.3.10                   
python                    3.8.19              
ultralytics               8.2.66               
